requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
openai>=1.0.0
duckdb>=0.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-docx>=0.8.11
python-multipart>=0.0.5
